+++ modules/model/src/main/java/net/pantasystem/milktea/model/user/UserDataSource.kt
@@ -3,29 +3,6 @@ package net.pantasystem.milktea.model.user
 import kotlinx.coroutines.flow.Flow
 import net.pantasystem.milktea.model.AddResult
 
-data class UsersState(
-    val usersMap: Map<User.Id, User> = emptyMap()
-) {
-
-    fun get(userId: User.Id?): User? {
-        return usersMap[userId]
-    }
-
-    fun get(userName: String, host: String? = null, accountId: Long? = null): User? {
-        return usersMap.values.filter {
-            accountId == null || accountId == it.id.accountId
-        }.firstOrNull {
-            it.userName == userName && it.host == host
-        }
-    }
-
-    fun get(accountId: Long, fqdnUserName: String): User? {
-        val userNameAndHost = fqdnUserName.split("@").filter { it.isNotBlank() }
-        val userName = userNameAndHost[0]
-        val host = userNameAndHost.getOrNull(1)
-        return get(userName, host, accountId)
-    }
-}
 
 interface UserDataSource {
